requirements.txt
Django==1.8
psycopg2==2.5.4

django-compressor==1.6

django-debug-toolbar==1.2.2

Markdown==2.6.5
django-filter==0.12.0
djangorestframework==3.3.2
